Return 0 instead of 0 in ovdec_receive_picture if DPB is not alloc yet
This might happen if the decoder only received non VCL NAL Units and we try
to see if the decoder has a picture to output.
This is not harmful to return 0 instead of an error here as long as
no operation on the DPB are attempted past this point
The DPB will be allocated later when receiving the first VCL NAL Unit.