Do not use pointer to const OVPicture in synchronisation function
While the picture should not be modified by those functions, the atomic_load of idx_function
requires non const access. Since it is part of OVPicture structure at the current time we
have to use a pointer to a non const OVPicture. This could be changed later by moving
thread synchronisation into another structure and referencing it in OVPicture